+++ changelog.md
@@ -2,9 +2,17 @@
 
 ## [Unreleased]
 
+## Added
+
+- Added support to import images in galleries from the plugin "Robo Gallery"
+
 ## Changes
 
 - Small optimizations for the multisite support
+
+### Fixed
+
+- Fixed missing hint for not existing synchronisation support for local external sources
 
 ## [5.0.3] - 18.04.2026